+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-for-patient.spec.tsx
@@ -61,7 +61,7 @@ describe('Patient dashboard for HCP', () => {
     })
 
     const secondaryHeader = await screen.findByTestId('patient-nav-bar')
-    expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsGenerate report')
+    expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsDownload report')
     await checkMedicalWidgetForPatient()
   })